# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SocialHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: getFeed
      tags: [posts]
      summary: Get personalized feed
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Feed posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostPage"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
  /posts/reels:
    get:
      operationId: getReels
      tags: [posts]
      summary: Get reels/short videos feed
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Reels
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostPage"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like or react to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Reaction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionResult"
  /posts/{postId}/save:
    post:
      operationId: savePost
      tags: [posts]
      summary: Save or unsave a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Save result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveResult"
  /posts/{postId}/share:
    post:
      operationId: sharePost
      tags: [posts]
      summary: Share a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Share result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareResult"
  /posts/{postId}/comments:
    get:
      operationId: getComments
      tags: [posts]
      summary: Get comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [posts]
      summary: Create a comment on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
  /posts/hashtag/{tag}:
    get:
      operationId: getPostsByHashtag
      tags: [posts]
      summary: Get posts by hashtag
      parameters:
        - name: tag
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Posts for hashtag
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  # ─── STORIES ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, username, displayName, createdAt]
      properties:
        id:
          type: string
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, creator, company, recruiter, influencer, moderator, admin]
        isVerified:
          type: boolean
        isPremium:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        isFollowing:
          type: boolean
        birthDate:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        experience:
          type: array
          items:
            $ref: '#/components/schemas/Experience'
        education:
          type: array
          items:
            $ref: '#/components/schemas/Education'
        languages:
          type: array
          items:
            $ref: '#/components/schemas/Language'
        socialLinks:
          type: array
          items:
            $ref: '#/components/schemas/SocialLink'
        createdAt:
          type: string
...
    Post:
      type: object
      required: [id, authorId, author, content, postType, createdAt]
      properties:
        id:
          type: string
        authorId:
          type: string
        author:
          $ref: "#/components/schemas/User"
        content:
          type: string
        postType:
          type: string
          enum: [text, image, video, audio, reel, poll, gif]
        mediaUrls:
          type: array
          items:
            type: string
        thumbnailUrl:
          type: ["string", "null"]
        hashtags:
          type: array
          items:
            type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        sharesCount:
          type: integer
        userReaction:
          type: ["string", "null"]
        isSaved:
          type: boolean
        isShared:
          type: boolean
        viewsCount:
          type: integer
        createdAt:
          type: string
    PostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        postType:
          type: string
          enum: [text, image, video, audio, reel, poll, gif]
        mediaUrls:
          type: array
          items:
            type: string
        hashtags:
          type: array
          items:
            type: string
    PostPage:
      type: object
      required: [posts]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        nextCursor:
          type: ["string", "null"]
        hasMore:
          type: boolean
    Comment:
      type: object
      required: [id, postId, authorId, author, content, createdAt]
      properties:
        id:
          type: string
        postId:
          type: string
        authorId:
          type: string
        author:
          $ref: "#/components/schemas/User"
        content:
          type: string
        likesCount:
          type: integer
        parentId:
          type: ["string", "null"]
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        parentId:
          type: string
    ReactionInput:
      type: object
      required: [reaction]
      properties:
        reaction:
          type: string
          enum: [like, love, haha, wow, sad, angry]
    ReactionResult:
      type: object
      required: [reaction, likesCount]
      properties:
        reaction:
          type: ["string", "null"]
        likesCount:
          type: integer
    SaveResult:
      type: object
      required: [saved]
      properties:
        saved:
          type: boolean
    ShareResult:
      type: object
      required: [shared, sharesCount]
      properties:
        shared:
          type: boolean
        sharesCount:
          type: integer

    # ─── STORY SCHEMAS ────────────────────────────────────────────────────────
...
    Skill:
      type: object
      required: [id, skill, level]
      properties:
        id:
          type: string
        skill:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
...
    Experience:
      type: object
      required: [id, company, title]
      properties:
        id:
          type: string
        company:
          type: string
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        current:
          type: boolean
...
    Education:
      type: object
      required: [id, institution, degree]
      properties:
        id:
          type: string
        institution:
          type: string
        degree:
          type: string
        field:
          type: string
        startYear:
          type: string
        endYear:
          type: string
        current:
          type: boolean
...
    Language:
      type: object
      required: [id, language, proficiency]
      properties:
        id:
          type: string
        language:
          type: string
        proficiency:
          type: string
          enum: [basic, conversational, fluent, native]
...
    SocialLink:
      type: object
      required: [id, platform, url]
      properties:
        id:
          type: string
        platform:
          type: string
        url:
          type: string
...