# lib/api-zod/src/generated/types/reactionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SocialHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReactionResult {
  /** @nullable */
  reaction: string | null;
  likesCount: number;
}